render movie rating

+++ src/Services/Rating/Rating.API/src/DTOs/RatingAnalysisDTO.cs
@@ -6,7 +6,7 @@ namespace IMBox.Services.Rating.API.DTOs
     {
         public MovieDTO Movie { get; init; }
         public IEnumerable<RatingItem> Ratings { get; init; }
-        public double AverageRating { get; init; }
+        public decimal AverageRating { get; init; }
         public int TotalRatingVoteCount { get; init; }
         public string DemographicType { get; init; }
